telegram version/requirements.txt
python-telegram-bot==20.7
Flask==2.3.3
python-dotenv==1.0.0
SQLAlchemy==2.0.23
aiohttp==3.9.1
Pillow==10.1.0
